+++ doc/source/changes.rst
@@ -3,6 +3,12 @@ Changelog
 #########
 
 **********
+v0.9.0
+**********
+- Fixed issue with resources never being freed as mmaps were never closed.
+- Client counting is now done manually, instead of relying on pyton's reference count
+
+**********
 v0.8.5
 **********
 - Fixed Python 3.0-3.3 regression, which also causes smmap to become about 3 times slower depending on the code path. It's related to this bug (http://bugs.python.org/issue15958), which was fixed in python 3.4